(From Sierra) stk_percept,adapt: more compile warning fixes

+++ packages/stk/stk_percept/stk_percept/Util.cpp
@@ -378,8 +378,8 @@ get_heap_info(
 
       ::getrusage(RUSAGE_SELF, &my_rusage);
 
-      double seconds = my_rusage.ru_utime.tv_sec + my_rusage.ru_stime.tv_sec;
-      double micro_seconds = my_rusage.ru_utime.tv_usec + my_rusage.ru_stime.tv_usec;
+      double seconds = (double)(my_rusage.ru_utime.tv_sec + my_rusage.ru_stime.tv_sec);
+      double micro_seconds = (double)(my_rusage.ru_utime.tv_usec + my_rusage.ru_stime.tv_usec);
       return seconds + micro_seconds*1.0e-6;
 
 #endif
